// apps/api/src/schemas/categories.ts
import z from "zod";

export const createCategorySchema = z.object({
  name: z.string().min(1, "Category name is required"),
  icon: z.string().optional(),
  color: z.string().optional(),
});
export type CreateCategory = z.infer<typeof createCategorySchema>;

export const updateCategorySchema = z.object({
  name: z.string().min(1, "Category name is required").optional(),
  icon: z.string().optional(),
  color: z.string().optional(),
});

export type UpdateCategory = z.infer<typeof updateCategorySchema>;